+++ Changelog.md
@@ -1,3 +1,8 @@
+# 1.2.15
+Correction d'anomalies sur le système de combat
+Ajout des dommaines Comprendre, Contrôler, Créer, Détruire et Transformer pour la magie analogique
+Ajout d'un onglet analogie pour les figures avec les domaines et les habitus
+
 # 1.2.14
 Correction d'une anomalie sur le calcul de l'initiative
 Correction d'une anomalie sur les tirs à distance